fix: reorganize crew code from streamlit app
Adopts familiar structure to current crewAI projects with a python file per type.
