--- app-v1/aibtcdev_config.yaml
@@ -1,28 +1,0 @@
-agents:
-- get_resource_manager
-- get_transaction_manager
-- get_wallet_manager
-app_settings:
-  layout: wide
-  page_title: AIBTCdev Crews
-file_settings:
-  agents_file: aibtcdev_agents.py
-  app_file: aibtcdev_app.py
-  config_file: aibtcdev_config.yaml
-  crews_file: aibtcdev_crews.py
-  tasks_file: aibtcdev_tasks.py
-  tools_file: aibtcdev_tools.py
-  utils_file: aibtcdev_utils.py
-model_settings:
-  Claude-Opus:
-    OPENAI_API_BASE: https://api.anthropic.com/v1
-    OPENAI_MODEL_NAME: claude-3-opus-20240229
-  Claude-Sonnet:
-    OPENAI_API_BASE: https://api.anthropic.com/v1
-    OPENAI_MODEL_NAME: claude-3-5-sonnet-20240620
-  GroqCloud:
-    OPENAI_API_BASE: https://api.groq.com/openai/v1
-    OPENAI_MODEL_NAME: Llama3-70b
-  OpenAI:
-    OPENAI_API_BASE: https://api.openai.com/v1
-    OPENAI_MODEL_NAME: test
