+++ app-v1/utils.py
@@ -8,6 +8,9 @@ from langchain_openai import ChatOpenAI
 CONFIG_FILENAME = "aibtcdev_config.yaml"
 
 
+# loads saved settings from config file
+# then injects environment variables into the config
+# if config file not found, generate by walking expected files
 def load_config():
     load_dotenv()
     with open(CONFIG_FILENAME, "r") as file:
